+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - EF Core performance monitoring ([#1112](https://github.com/getsentry/sentry-dotnet/pull/1112))
 
+## Fixes
+
+- Fix max breadcrumbs limit when MaxBreadcrumbs is zero or lower ([#1145](https://github.com/getsentry/sentry-dotnet/pull/1145))
+
 ## 3.8.3
 
 ### Features